__str__ refactor. Ellipse CRs added. Mapping repr refactor.
Refactored to have AbstractGeometry create its own __str__ prefix that can be used by its subclasses. Added a verbosity flag to allow it to add the uid of the geometry. Added Ellipse ConstraintReferences for each extreme point along with tests for them. Added constraint mapping for freecad and added the ability to print their mapping to the command line.