fixed constraint map reference bug and exported first constrained ellipse.
Corrected the constraint map assign internal constraints to set the geometry index and not the constraint index. Added self to ellipse get_reference. Corrected ORIGIN reference to the start subpart instead of edge. Added constraints to the ellipse sample sketch. Added Ellipse to the validation list of each constraint type.