deleted deprecated freecad code and renamed files to private.
Deleted freecad translation code that's out of date now that there is a new mapping system. Renamed all internal freecad modules to private and reduced the public import info to just the FreeCADFile and the App, Part, Sketcher modules. Moved freecad code imports that were just for typing into an if statement for type checking