147 freecad pointonobject (#151)
* added PointOnObject write ability and tested with square variation

* added the ability to read in PointOnObject and added tests.

Added new FCStd to act as a challenge model for the freecad map. Added pointonobject tangent translation for line to line coincidents in addition to having point on line coincident. Refactored ConstraintType to be able to return the Type of the freecad constraints in those situations by adding the constraint as an argument to its translator methods.